Further work on LDAP logins.
LDAP accounts are automatically created on first login, and populate given name, family name, and email by reading the LDAP directory.
This is hardcoded within the portal to work with the University of Melbourne's "centaur" system.
LDAP accounts ought not to have an entry in the membership table, but unfortunately the creation date is stored there rather than in the user table, so it can't be null without breaking the existing code.  I'm not willing to tinker with the database scheme right now to correct that.