Initial commit of the Chiaroscuro ZK theme.
Chiaroscuro uses CSS variables from a central CSS file rather than hardcoding colors and requiring the entire theme to be changed -- it's "the last ZK theme you'll ever need".
Because compiling it requires lessc, none of this is built by default, and there isn't a profile to activate it yet, either.
The portal and process discoverer are correctly linked to the new CSS the Chiaroscuro requires.
The zk-osgi bundle DOESN'T include the Chiaroscuro theme yet.