Enable building on JDK 11+
The primary thing added here is a profile in the top-level pom.xml to add various packages which were removed from Java SE in version 11.
A secondary change is that Test-Tests now inherits the compiler settings from the top-level pom.xml, with the side effect that the license plugin has reformatted all its source code.